refactor(memory): remove memorizer and file-backed memory backends
Deletes the Memorizer-MCP and file-backed tool implementations, their
tests, and all conditional branches that routed to them. SQLite is the
only supported memory backend.

Changes:
- Delete 11 source files: Memorizer*Tool, MemorizerMemoryExtractor,
  File*Tool, FileMemoryStore, FileMemoryExtractor, StoreMemoryTool
- Delete 12 test files covering the removed tools
- Remove FileBacked/MemorizerConnected/MemorizerDisconnected from
  MemoryContextState enum and MemoryIndexContextLayer switch
- Remove MemoryConfig.Provider property and schema enum (sqlite only)
- Simplify ToolIndexUpdater: drop _memoryConfig field, hardcode
  ResolveMemoryState() to SqlitePrimary
- Simplify DaemonRuntimeStatusService: drop fileMemoryStore param and
  memorizer/file branches from BuildMemoryStatusAsync
- Remove memorizer-usage system skill and embedded copy
- Remove memorizer references from capability-reference skill
- Extract FakeNetclawTool to shared test helper (was in deleted file)
- Update ConfigSchemaDoctorCheckTests to not reference removed Provider
