+++ src/Netclaw.Configuration/MemoryConfig.cs
@@ -2,15 +2,10 @@ namespace Netclaw.Configuration;
 
 /// <summary>
 /// Configuration for the cross-session memory subsystem.
-/// SQLite-first durable memory settings.
+/// SQLite-backed durable memory settings.
 /// </summary>
 public sealed class MemoryConfig
 {
-    /// <summary>
-    /// Durability backend. MVP defaults to local SQLite.
-    /// </summary>
-    public string Provider { get; set; } = "sqlite";
-
     /// <summary>
     /// Automatic recall timeout budget in milliseconds.
     /// </summary>
