+++ src/Netclaw.Daemon.Tests/Gateway/DaemonRuntimeStatusServiceTests.cs
@@ -29,8 +29,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
         Main = new ModelReference { Provider = "test-provider", ModelId = "test-model" }
     };
 
-    private static readonly MemoryConfig DefaultMemoryConfig = new();
-
     private readonly string _tempBase = Path.Combine(Path.GetTempPath(), $"netclaw-status-test-{Guid.NewGuid():N}");
 
     private NetclawPaths CreatePaths() => new(_tempBase);
@@ -81,7 +79,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
             telemetryOptions: Options.Create(new TelemetryOptions()),
             sessionConfig: DefaultSessionConfig,
             modelSelection: DefaultModelSelection,
-            memoryConfig: DefaultMemoryConfig,
             paths: CreatePaths());
 
         var status = await service.GetStatusAsync();
@@ -102,7 +99,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
             telemetryOptions: Options.Create(new TelemetryOptions()),
             sessionConfig: DefaultSessionConfig,
             modelSelection: DefaultModelSelection,
-            memoryConfig: DefaultMemoryConfig,
             paths: CreatePaths());
 
         var status = await service.GetStatusAsync();
@@ -123,7 +119,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
             telemetryOptions: Options.Create(new TelemetryOptions()),
             sessionConfig: DefaultSessionConfig,
             modelSelection: DefaultModelSelection,
-            memoryConfig: DefaultMemoryConfig,
             paths: CreatePaths());
 
         var status = await service.GetStatusAsync();
@@ -178,7 +173,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
                 telemetryOptions: Options.Create(new TelemetryOptions()),
                 sessionConfig: DefaultSessionConfig,
                 modelSelection: DefaultModelSelection,
-                memoryConfig: DefaultMemoryConfig,
                 paths: CreatePaths(),
                 mcpClientManager: manager);
 
@@ -200,58 +194,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
     }
 
     [Fact]
-    public async Task StatusIncludesMemory_FileBackend()
-    {
-        var paths = CreatePaths();
-        paths.EnsureDirectoriesExist();
-        var fileStore = new FileMemoryStore(paths.MemoriesDirectory, TimeProvider.System);
-
-        await fileStore.StoreAsync("First Memory", "Content one.");
-        await fileStore.StoreAsync("Second Memory", "Content two.");
-
-        var service = new DaemonRuntimeStatusService(
-            TimeProvider.System,
-            channels: Array.Empty<IChannel>(),
-            slackOptions: new SlackChannelOptions { Enabled = false },
-            persistenceOptions: new DaemonPersistenceOptions(),
-            telemetryOptions: Options.Create(new TelemetryOptions()),
-            sessionConfig: DefaultSessionConfig,
-            modelSelection: DefaultModelSelection,
-            memoryConfig: new MemoryConfig { Provider = "files" },
-            paths: paths,
-            fileMemoryStore: fileStore);
-
-        var status = await service.GetStatusAsync();
-
-        Assert.NotNull(status.Memory);
-        Assert.Equal("files", status.Memory.Provider);
-        Assert.Equal("healthy", status.Memory.Status);
-        Assert.Equal(2, status.Memory.MemoryCount);
-        Assert.Equal(paths.MemoryIndexPath, status.Memory.IndexPath);
-    }
-
-    [Fact]
-    public async Task StatusIncludesMemory_MemorizerNotConnected()
-    {
-        var service = new DaemonRuntimeStatusService(
-            TimeProvider.System,
-            channels: Array.Empty<IChannel>(),
-            slackOptions: new SlackChannelOptions { Enabled = false },
-            persistenceOptions: new DaemonPersistenceOptions(),
-            telemetryOptions: Options.Create(new TelemetryOptions()),
-            sessionConfig: DefaultSessionConfig,
-            modelSelection: DefaultModelSelection,
-            memoryConfig: new MemoryConfig { Provider = "memorizer" },
-            paths: CreatePaths());
-
-        var status = await service.GetStatusAsync();
-
-        Assert.NotNull(status.Memory);
-        Assert.Equal("memorizer", status.Memory.Provider);
-        Assert.Equal("unavailable", status.Memory.Status);
-    }
-
-    [Fact]
     public async Task StatusIncludesMemory_SqliteBackend()
     {
         var paths = CreatePaths();
@@ -268,7 +210,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
             telemetryOptions: Options.Create(new TelemetryOptions()),
             sessionConfig: DefaultSessionConfig,
             modelSelection: DefaultModelSelection,
-            memoryConfig: new MemoryConfig { Provider = "sqlite" },
             paths: paths,
             sqliteMemoryStore: sqliteStore);
 
@@ -301,7 +242,6 @@ public sealed class DaemonRuntimeStatusServiceTests : IAsyncLifetime
             telemetryOptions: Options.Create(new TelemetryOptions()),
             sessionConfig: DefaultSessionConfig,
             modelSelection: DefaultModelSelection,
-            memoryConfig: DefaultMemoryConfig,
             paths: CreatePaths());
 
         var status = await service.GetStatusAsync();
